service/requirements.txt
asgiref==3.7.2
class-registry==2.1.2
coverage==7.3.2
Django==4.2.6
django-cors-headers==4.3.0
django-filter==23.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
filters==1.3.2
iniconfig==2.0.0
Jinja2==3.1.2
MarkupSafe==2.1.3
mypy==1.6.1
mypy-extensions==1.0.0
packaging==23.2
Pillow==10.1.0
pluggy==1.3.0
PyJWT==2.8.0
pytest==7.4.3
pytest-cov==4.1.0
pytest-django==4.6.0
pytest-html==4.0.2
pytest-metadata==3.0.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
regex==2023.10.3
setuptools==68.0.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.8.0
wheel==0.37.1
